Final Version
This uses OOP to recreate the game Pong, I used the Turtle module to be able to draw the paddles, the ball, and score board . We have the main.py which contains the sequence of the code that runs all the classes and the sequence for the game pong. The ball contains the constructor for the ball and how it resets and bounces the wall and the paddles. We also have the paddle class, where we have the method of up and down that controls the way the paddles move, and lastly the scoreboard which  keep tracks and display the score